Retry db commit if db locked on attempt.
If anything else is accessing the database when eddblink tries to modify
it, it would crash with a database is locked error.

Now if it gets that error, it will wait 1 second and then try again
until it succeeds rather than crashing if it fails.

Note that this has no effect on any other errors that may occur.