Don't DL live listings from fallback, unmark as live when dump updated.
The file download method needs to quit when being told to download the
live listings from the fallback, not merely when it can't access them
from the mirror, because the 'fallback' option might be turned on.

The EDDB dump will never have data more recent than the data from an
active listener, but it will have data /as/ recent. When that happens,
we want to reset the field in the database that says the data is from
the listener's updates, that way the live listings doesn't duplicate
data found in the dump.