Run with 'listings' enabled by default.
If no options have been passed, enable listings as a default.

The options 'force' and 'fallback' are excluded from this, as they don't
say what to import so much as how.

The option 'skipvend' is excluded because it says what /not/ to import.

If no options other than these are passed, listings will be enabled. If
any of the others are passed, the plugin will proceed as normal.